+++ src/app/component/detail/Chart/Financial.tsx
