+++ src/app/component/detail/Chart/Pie.tsx
