+++ src/app/component/detail/Initialize.tsx
