+++ src/app/component/detail/StockSearch.tsx
